+++ chopper/example/main.dart
@@ -1,4 +1,7 @@
+import 'dart:async';
+
 import 'package:chopper/chopper.dart';
+import 'package:chopper/src/interceptor.dart';
 import 'definition.dart';
 
 main() async {
@@ -8,7 +11,7 @@ main() async {
       // the generated service
       MyService.create()
     ],
-    jsonApi: true,
+    converter: JsonConverter(),
   );
 
   final myService = chopper.service<MyService>(MyService);
@@ -16,5 +19,23 @@ main() async {
   final response = await myService.getMapResource("1");
   print(response.body);
 
-  chopper.close();
+  final list = await myService.getListResources();
+  print(list.body);
+
+  chopper.dispose();
+}
+
+class _Converter implements Converter {
+  @override
+  FutureOr<Request> convertRequest(Request request) {
+    // TODO: implement convertRequest
+    return null;
+  }
+
+  @override
+  FutureOr<Response<ConvertedResponseType>>
+      convertResponse<ConvertedResponseType>(Response response) {
+    // TODO: implement convertResponse
+    return null;
+  }
 }