+++ chopper/lib/src/base.dart
@@ -1,16 +1,14 @@
 import "dart:async";
-import 'dart:convert' hide Converter;
-import 'package:chopper/src/interceptor.dart';
 import "package:meta/meta.dart";
 import 'package:http/http.dart' as http;
 
 import "interceptor.dart";
 import "request.dart";
 import 'response.dart';
+import 'annotations.dart';
 
 /// Root object of chopper
 /// Used to manager services, encode data, intercept request, response and error.
-@immutable
 class ChopperClient {
   /// base url of each request to your api
   /// hostname of your api for example
@@ -27,20 +25,14 @@ class ChopperClient {
   /// converter call on error request
   final Converter errorConverter;
 
-  /// default: false
-  /// Use to define your a api need json encoded data
-  final bool jsonApi;
-
-  /// default: true
-  /// Use to define your a api need form url encoded data
-  final bool formUrlEncodedApi;
-
   final Map<Type, ChopperService> _services = {};
   final _requestInterceptors = [];
   final _responseInterceptors = [];
   final _requestController = StreamController<Request>.broadcast();
   final _responseController = StreamController<Response>.broadcast();
   final _responseErrorController = StreamController<Response>.broadcast();
+
+  final bool _clientIsInternal;
 
   ChopperClient({
     this.baseUrl: "",
@@ -49,9 +41,8 @@ class ChopperClient {
     this.converter,
     this.errorConverter,
     Iterable<ChopperService> services: const [],
-    this.jsonApi: false,
-    this.formUrlEncodedApi: true,
-  }) : httpClient = client ?? http.Client() {
+  })  : httpClient = client ?? http.Client(),
+        _clientIsInternal = client == null {
     if (interceptors.every(_isAnInterceptor) == false) {
       throw Exception(
           "Unsupported type for interceptors, it only support the following types: RequestInterceptor, RequestInterceptorFunc, ResponseInterceptor, ResponseInterceptorFunc");
@@ -61,9 +52,8 @@ class ChopperClient {
     _responseInterceptors.addAll(interceptors.where(_isResponseInterceptor));
 
     services.toSet().forEach((s) {
-      final mix = (s as ChopperServiceMixin);
-      mix.client = this;
-      _services[mix.definitionType] = s;
+      s.client = this;
+      _services[s.definitionType] = s;
     });
   }
 
@@ -85,26 +75,11 @@ class ChopperClient {
   }
 
   Future<Request> _encodeRequest(Request request) async {
-    final convertJson = jsonApi || request.json == true;
-
-    var converted = request;
-
     if (converter != null) {
-      converted = await converter.encode(request);
+      return converter.convertRequest(request);
     }
 
-    if (convertJson) {
-      if (converted.body != null) {
-        return converted.replace(body: json.encode(converted.body));
-      } else if (converted.parts?.isNotEmpty == true) {
-        final parts = converted.parts
-            .map((p) => p.replace(value: json.encode(p.value)))
-            .toList();
-        return converted.replace(parts: parts);
-      }
-    }
-
-    return converted;
+    return request;
   }
 
   Future<Response> _decodeResponse<Value>(
@@ -113,7 +88,7 @@ class ChopperClient {
   ) async {
     if (withConverter == null) return response;
 
-    final converted = await withConverter.decode<Value>(response);
+    final converted = await withConverter.convertResponse<Value>(response);
 
     if (converted == null) {
       throw Exception("No converter found for type $Value");
@@ -146,32 +121,35 @@ class ChopperClient {
     return res;
   }
 
-  Future<Response<Value>> send<Value>(Request request) async {
+  Future<Response> send<Value>(
+    Request request, {
+    ConvertRequest requestConverter,
+    ConvertResponse<Value> responseConverter,
+  }) async {
     Request req = request;
 
     if (req.body != null || req.parts.isNotEmpty) {
-      req = await _encodeRequest(request);
+      if (requestConverter != null) {
+        req = await requestConverter(request);
+      } else {
+        req = await _encodeRequest(request);
+      }
     }
 
     req = await _interceptRequest(req);
 
     _requestController.add(req);
-    final stream = await httpClient.send(await req.toHttpRequest(
-      baseUrl,
-      jsonApi: jsonApi,
-      formUrlEncodedApi: formUrlEncodedApi,
-    ));
+    final stream = await httpClient.send(await req.toHttpRequest());
 
     final response = await http.Response.fromStream(stream);
-
     Response res = Response(response, response.body);
 
-    if ((jsonApi == true && req.formUrlEncoded != true) || req.json == true) {
-      res = _tryDecodeJson(res);
-    }
-
     if (res.isSuccessful) {
-      res = await _decodeResponse<Value>(res, converter);
+      if (responseConverter != null) {
+        res = await responseConverter(res);
+      } else {
+        res = await _decodeResponse<Value>(res, converter);
+      }
     } else {
       res = await _decodeResponse(res, errorConverter);
     }
@@ -187,33 +165,128 @@ class ChopperClient {
     return res;
   }
 
-  Response _tryDecodeJson(Response res) {
-    try {
-      return res.replace(body: json.decode(res.body));
-    } catch (_) {
-      return res;
-    }
+  Future<Response<Body>> get<Body>(
+    String url, {
+    Map<String, String> headers,
+  }) =>
+      send(
+        Request(
+          'GET',
+          url,
+          baseUrl,
+          headers: headers,
+        ),
+      );
+
+  Future<Response<Body>> post<Body>(
+    String url, {
+    dynamic body,
+    List<PartValue> parts,
+    Map<String, String> headers,
+  }) =>
+      send(
+        Request(
+          'POST',
+          url,
+          baseUrl,
+          body: body,
+          parts: parts,
+          headers: headers,
+        ),
+      );
+
+  Future<Response<Body>> put<Body>(
+    String url, {
+    dynamic body,
+    List<PartValue> parts,
+    Map<String, String> headers,
+  }) =>
+      send(
+        Request(
+          'PUT',
+          url,
+          baseUrl,
+          body: body,
+          parts: parts,
+          headers: headers,
+        ),
+      );
+
+  Future<Response<Body>> patch<Body>(
+    String url, {
+    dynamic body,
+    List<PartValue> parts,
+    Map<String, String> headers,
+  }) =>
+      send(
+        Request(
+          'PATCH',
+          url,
+          baseUrl,
+          body: body,
+          parts: parts,
+          headers: headers,
+        ),
+      );
+
+  Future<Response<Body>> delete<Body>(
+    String url, {
+    Map<String, String> headers,
+  }) =>
+      send(
+        Request(
+          'DELETE',
+          url,
+          baseUrl,
+          headers: headers,
+        ),
+      );
+
+  @Deprecated('use dispose')
+  @mustCallSuper
+  void close() {
+    dispose();
   }
 
   /// dispose [ChopperClient] to clean memory
-  void close() {
+  @mustCallSuper
+  void dispose() {
     _requestController.close();
     _responseController.close();
     _responseErrorController.close();
+
+    _services.forEach((_, s) => s.dispose());
+    _services.clear();
+
+    _requestInterceptors.clear();
+    _responseInterceptors.clear();
+
+    if (_clientIsInternal) {
+      httpClient.close();
+    }
   }
 
+  /// Event stream of request just before http call
+  /// all converters and interceptors have been run
   Stream<Request> get onRequest => _requestController.stream;
 
+  /// Event stream of response
+  /// all converters and interceptors have been run
   Stream<Response> get onResponse => _responseController.stream;
 
+  /// Event stream of error response (status code <200 >=300)
+  /// all converters and interceptors have been run
   Stream<Response> get onError => _responseErrorController.stream;
 }
 
 /// Used by generator to generate apis
-abstract class ChopperService {}
-
-abstract class ChopperServiceMixin {
+abstract class ChopperService {
   ChopperClient client;
 
-  Type get definitionType;
+  Type get definitionType => null;
+
+  @mustCallSuper
+  void dispose() {
+    client = null;
+  }
 }