+++ chopper/test/form_test.dart
@@ -11,17 +11,15 @@ void main() {
     };
 
     group('Form', () {
-      final buildClient = (bool form, http.Client httpClient,
-              {bool json: false}) =>
-          ChopperClient(
-            services: [
-              // the generated service
-              HttpTestService.create(),
-            ],
-            client: httpClient,
-            jsonApi: json,
-            formUrlEncodedApi: form,
-          );
+      final buildClient =
+          (http.Client httpClient, {bool json: false}) => ChopperClient(
+                services: [
+                  // the generated service
+                  HttpTestService.create(),
+                ],
+                client: httpClient,
+                converter: FormUrlEncodedConverter(),
+              );
 
       test('form-urlencoded default', () async {
         final httpClient = MockClient((http.Request req) async {
@@ -34,33 +32,11 @@ void main() {
           return http.Response('ok', 200);
         });
 
-        final chopper = buildClient(false, httpClient);
+        final chopper = buildClient(httpClient);
 
         final result = await chopper
             .service<HttpTestService>(HttpTestService)
             .mapTest(sample);
-
-        expect(result.body, equals('ok'));
-
-        httpClient.close();
-      });
-
-      test('force form-urlencoded', () async {
-        final httpClient = MockClient((http.Request req) async {
-          expect(req.url.toString(), equals('/test/map/form'));
-          expect(
-            req.headers['content-type'],
-            'application/x-www-form-urlencoded; charset=utf-8',
-          );
-          expect(req.body, 'foo=bar');
-          return http.Response('ok', 200);
-        });
-
-        final chopper = buildClient(false, httpClient);
-
-        final result = await chopper
-            .service<HttpTestService>(HttpTestService)
-            .forceFormTest(sample);
 
         expect(result.body, equals('ok'));