+++ chopper/test/multipart_test.dart
@@ -11,29 +11,30 @@ void main() {
         expect(req.headers['Content-Type'], contains('multipart/form-data;'));
         expect(
           req.body,
-          contains('''
-content-disposition: form-data; name="1"\r
-\r
-{"foo":"bar"}\r
-'''),
+          contains(
+            'content-disposition: form-data; name="1"\r\n'
+                '\r\n'
+                '{foo: bar}\r\n',
+          ),
         );
         expect(
           req.body,
-          contains('''
-content-disposition: form-data; name="2"\r
-\r
-{"bar":"foo"}\r
-'''),
+          contains(
+            'content-disposition: form-data; name="2"\r\n'
+                '\r\n'
+                '{bar: foo}\r\n',
+          ),
         );
         return http.Response('ok', 200);
       });
 
-      final chopper = ChopperClient(client: httpClient, jsonApi: true);
+      final chopper =
+          ChopperClient(client: httpClient, converter: JsonConverter());
       final service = HttpTestService.create(chopper);
 
       await service.postResources({'foo': 'bar'}, {'bar': 'foo'});
 
-      chopper.close();
+      chopper.dispose();
     });
 
     test('file', () async {
@@ -59,7 +60,7 @@ ${String.fromCharCodes([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])}\r
 
       await service.postFile([0, 1, 2, 3, 4, 5, 6, 7, 8, 9]);
 
-      chopper.close();
+      chopper.dispose();
     });
   });
 }