+++ chopper/test/test_service.dart
@@ -1,4 +1,5 @@
 import "dart:async";
+import 'dart:convert';
 import 'package:chopper/chopper.dart';
 
 part "test_service.chopper.dart";
@@ -8,45 +9,53 @@ abstract class HttpTestService extends ChopperService {
   static HttpTestService create([ChopperClient client]) =>
       _$HttpTestService(client);
 
-  @Get(url: "get/{id}")
+  @Get(path: "get/{id}")
   Future<Response> getTest(
     @Path() String id, {
     @Header('test') String dynamicHeader,
   });
 
-  @Post(url: "post")
+  @Post(path: "post")
   Future<Response> postTest(@Body() String data);
 
-  @Put(url: 'put/{id}')
+  @Put(path: 'put/{id}')
   Future<Response> putTest(@Path('id') String test, @Body() String data);
 
-  @Delete(url: 'delete/{id}', headers: {'foo': 'bar'})
+  @Delete(path: 'delete/{id}', headers: {'foo': 'bar'})
   Future<Response> deleteTest(@Path() String id);
 
-  @Patch(url: 'patch/{id}')
+  @Patch(path: 'patch/{id}')
   Future<Response> patchTest(@Path() String id, @Body() String data);
 
-  @Post(url: 'map')
+  @Post(path: 'map')
   Future<Response> mapTest(@Body() Map map);
 
-  @Post(url: 'map/json')
-  @JsonEncoded()
+  @Post(path: 'map/json')
+  @FactoryConverter(
+      request: customConvertRequest, response: customConvertResponse)
   Future<Response> forceJsonTest(@Body() Map map);
 
-  @Post(url: 'map/form')
-  @FormUrlEncoded()
-  Future<Response> forceFormTest(@Body() Map map);
-
-  @Post(url: 'multi')
+  @Post(path: 'multi')
   @multipart
   Future<Response> postResources(
     @Part('1') Map a,
     @Part('2') Map b,
   );
 
-  @Post(url: 'file')
+  @Post(path: 'file')
   @multipart
   Future<Response> postFile(
     @FileField('file') List<int> bytes,
   );
+
+  @Get(path: 'https://test.com')
+  Future fullUrl();
 }
+
+Request customConvertRequest(Request req) {
+  final r = JsonConverter().convertRequest(req);
+  return applyHeader(r, 'customConverter', 'true');
+}
+
+Response<T> customConvertResponse<T>(Response res) =>
+    res.replace(body: json.decode(res.body));