+++ chopper_generator/lib/src/generator.dart
@@ -12,14 +12,14 @@ import 'package:source_gen/source_gen.dart';
 import 'package:code_builder/code_builder.dart';
 import 'package:chopper/chopper.dart' as chopper;
 
-const _urlVar = "url";
-const _baseUrlVar = "baseUrl";
-const _parametersVar = "params";
-const _headersVar = "headers";
-const _requestVar = "request";
-const _bodyVar = 'body';
-const _partsVar = 'parts';
 const _clientVar = 'client';
+const _baseUrlVar = "baseUrl";
+const _parametersVar = "\$params";
+const _headersVar = "\$headers";
+const _requestVar = "\$request";
+const _bodyVar = '\$body';
+const _partsVar = '\$parts';
+const _urlVar = "\$url";
 
 class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
   @override
@@ -73,8 +73,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         ])
         ..methods.addAll(_parseMethods(element, baseUrl))
         ..fields.add(_buildDefinitionTypeMethod(friendlyName))
-        ..extend = refer(friendlyName)
-        ..mixins.add(refer('${chopper.ChopperServiceMixin}'));
+        ..extend = refer(friendlyName);
     });
 
     final emitter = new DartEmitter();
@@ -103,8 +102,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
   Method _generateMethod(MethodElement m, String baseUrl) {
     final method = _getMethodAnnotation(m);
     final multipart = _hasAnnotation(m, chopper.Multipart);
-    final formUrlEncoded = _hasAnnotation(m, chopper.FormUrlEncoded);
-    final hasJson = _hasAnnotation(m, chopper.JsonEncoded);
+    final factoryConverter = _getFactoryConverterAnotation(m);
 
     final body = _getAnnotation(m, chopper.Body);
     final paths = _getAnnotations(m, chopper.Path);
@@ -151,8 +149,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         blocks.add(headers);
       }
 
-      final hasBody =
-          body.isNotEmpty || (formUrlEncoded == true && fields.isNotEmpty);
+      final hasBody = body.isNotEmpty || fields.isNotEmpty;
       if (hasBody) {
         if (body.isNotEmpty) {
           blocks.add(
@@ -178,11 +175,26 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         useQueries: queries.isNotEmpty,
         useHeaders: headers != null,
         hasParts: hasParts,
-        hasFormUrlEncoded: formUrlEncoded,
-        hasJson: hasJson,
       ).assignFinal(_requestVar).statement);
 
       final namedArguments = <String, Expression>{};
+
+      final requestFactory = factoryConverter?.peek('request');
+      if (requestFactory != null) {
+        final el = requestFactory.objectValue.type.element;
+        if (el is FunctionTypedElement) {
+          namedArguments['requestConverter'] = refer(_factoryForFunction(el));
+        }
+      }
+
+      final responseFactory = factoryConverter?.peek('response');
+      if (responseFactory != null) {
+        final el = responseFactory.objectValue.type.element;
+        if (el is FunctionTypedElement) {
+          namedArguments['responseConverter'] = refer(_factoryForFunction(el));
+        }
+      }
+
       final typeArguments = <Reference>[];
       if (responseType != null) {
         typeArguments.add(refer(responseType.displayName));
@@ -195,6 +207,13 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
 
       b.body = new Block.of(blocks);
     });
+  }
+
+  String _factoryForFunction(FunctionTypedElement function) {
+    if (function.enclosingElement is ClassElement) {
+      return '${function.enclosingElement.name}.${function.name}';
+    }
+    return function.name;
   }
 
   Map<String, ConstantReader> _getAnnotation(MethodElement m, Type type) {
@@ -236,6 +255,13 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     return null;
   }
 
+  ConstantReader _getFactoryConverterAnotation(MethodElement method) {
+    final annot = _typeChecker(chopper.FactoryConverter)
+        .firstAnnotationOf(method, throwOnUnresolved: false);
+    if (annot != null) return new ConstantReader(annot);
+    return null;
+  }
+
   bool _hasAnnotation(MethodElement method, Type type) {
     final annot =
         _typeChecker(type).firstAnnotationOf(method, throwOnUnresolved: false);
@@ -260,14 +286,13 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
 
   DartType _getResponseType(DartType type) {
     final generic = _genericOf(type);
-    if (generic == null ||
-        _typeChecker(Map).isExactlyType(type) ||
-        _typeChecker(List).isExactlyType(type)) {
-      return type;
-    }
-    if (generic.isDynamic) {
-      return null;
-    }
+
+    if (generic == null || _typeChecker(Map).isExactlyType(type)) return type;
+
+    if (generic.isDynamic) return null;
+
+    if (_typeChecker(List).isExactlyType(type)) return generic;
+
     return _getResponseType(generic);
   }
 
@@ -276,16 +301,23 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     Map<ParameterElement, ConstantReader> paths,
     String baseUrl,
   ) {
-    String value = "${method.read("url").stringValue}";
+    String value = "${method.read("path").stringValue}";
     paths.forEach((p, ConstantReader r) {
       final name = r.peek("name")?.stringValue ?? p.displayName;
       value = value.replaceFirst("{$name}", "\$${p.displayName}");
     });
-    if (!baseUrl.endsWith('/') && !value.startsWith('/')) {
-      return literal('$baseUrl/$value');
-    }
 
-    return literal('$baseUrl$value');
+    if (value.startsWith('http://') || value.startsWith('https://')) {
+      // if the request's url is already a fully qualified URL, we can use
+      // as-is and ignore the baseUrl
+      return literal(value);
+    } else {
+      if (!baseUrl.endsWith('/') && !value.startsWith('/')) {
+        return literal('$baseUrl/$value');
+      }
+
+      return literal('$baseUrl$value');
+    }
   }
 
   Expression _generateRequest(
@@ -294,29 +326,22 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     bool hasParts: false,
     bool useQueries: false,
     bool useHeaders: false,
-    bool hasFormUrlEncoded: false,
-    bool hasJson: false,
   }) {
     final params = <Expression>[
       literal(method.peek("method").stringValue),
-      refer(_urlVar)
+      refer(_urlVar),
+      refer('$_clientVar.$_baseUrlVar'),
     ];
 
     final namedParams = <String, Expression>{};
 
     if (hasBody) {
-      namedParams[_bodyVar] = refer(_bodyVar);
+      namedParams['body'] = refer(_bodyVar);
     }
 
     if (hasParts) {
-      namedParams[_partsVar] = refer(_partsVar);
+      namedParams['parts'] = refer(_partsVar);
       namedParams['multipart'] = literalBool(true);
-    }
-
-    if (hasJson) {
-      namedParams['json'] = literalBool(true);
-    } else if (hasFormUrlEncoded) {
-      namedParams['formUrlEncoded'] = literal(true);
     }
 
     if (useQueries) {